Reorganise files & rename
Change Log:
- Relocated all the components.
  : All the related components are gathered at the same folder.
- Renamed components according to vue style guide
  : All the components should have multi-word name
    to prevent conflicts with existing and future HTML elements.
- Changed the way importing components (router)
  : Importing components should be checked in compile time, not in runtime.
