+++ .env.development
@@ -1,2 +1,2 @@
 NODE_ENV=development
-VUE_APP_BASE_API=https://3.34.46.127/
+VUE_APP_BASE_API=https://ops-test.com/
