+++ src/components/common/ResizeComponent.vue
