+++ src/components/layout/LayoutComponent.vue
@@ -6,24 +6,21 @@
         <router-view></router-view>
       </v-container>
     </v-content>
-    <Footer />
+    <FooterComponent />
   </v-app>
 </template>
 
 <script lang="ts">
 import Vue from "vue";
 import Component from "vue-class-component";
-import AppBar from "@/components/layout/AppBar.vue";
-import SideBar from "@/components/layout/SideBar.vue";
-import Footer from "@/components/layout/Footer.vue";
+import AppBar from "@/components/layout/app-bar/AppBar.vue";
+import FooterComponent from "@/components/layout/FooterComponent.vue";
 
 @Component({
   components: {
     AppBar,
-    SideBar,
-    Footer
+    FooterComponent
   }
 })
 export default class BasicLayout extends Vue {}
 </script>
-<style lang="scss" scoped></style>
