+++ src/components/main/PostCardView.vue
