+++ src/components/main/sidebar/Categories.vue
