+++ src/main.ts
@@ -3,13 +3,13 @@ import App from "@/App.vue";
 import router from "@/router";
 import store from "@/store";
 import vuetify from "@/plugins/vuetify";
-import Resize from "@/components/common/Resize.vue";
-import Modal from "@/components/common/Modal.vue";
+import ResizeComponent from "@/components/common/ResizeComponent.vue";
+import ModalComponent from "@/components/common/ModalComponent.vue";
 import "@/utils/class-component-hooks";
 
 Vue.config.productionTip = false;
-Vue.component("v-resize", Resize);
-Vue.component("v-modal", Modal);
+Vue.component("v-resize", ResizeComponent);
+Vue.component("v-modal", ModalComponent);
 
 new Vue({
   router,
